requirements.txt
peewee==2.10.2
discord.py==1.0.0a1650+g1222bce
feedparser==5.2.1
aiohttp==3.4.4
GitPython==2.1.11